Initial commit
This is a library that works with E-Series¹. Currently, it implements the
series in the common variants E3, E6, E12, E24, E48, E96 and E192. It also
implements a set of functions to lookup adjacent values to arbitrary values
from these series. The plan is to add functions to generate parallel and se-
ries combinations of E-Series values to match arbitrary values as closely as
possible.

¹ https://en.wikipedia.org/wiki/E_series_of_preferred_numbers

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>