Use a work-around to make library work on Guile 2.0
It seems like, if you have a module

  (foo)

and a module

  (foo bar)

with the latter defining a procedure "bar" and you're importing the latter in
the former, then within (foo) evaluating "bar" will result in a reference to
the module (foo bar). And with that:

scheme@(e-series)> combine
$1 = #<directory (e-series combine) 55def04e8120>
scheme@(e-series)> (module? combine)
$2 = #t

It's key that (foo bar) is a sub-module of (foo), namespace-wise. If you have a
module

  (foo baz)

that imports (foo bar), it can use "bar" just like that; in Guile 2.0 as well.

The work-around is to import modules like that with symbols to be prefixed by a
dollar sign ($). Then use procedure by names like "$combine" instead of the
original "combine".

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>